GUI update + some backend
Even though it was planned to focus on backend stuff first, it becomes increasingly difficult to imagine how the user can interact with it without a working GUI at the spot. As such, I determined that they should be developed in parallel.

So far, you can login as the default admin, and log out with ease. No real actions have been implemented yet, however.