Another GUI update w/ lil backend
For the most part, I think the option select screen for each kind of User is now done... ish. 
The real pain now is going to be implementing the GUI and backend necessary to complete option available for each user.

Once that is out of the way, however, the main base of the project would be done.