Print function + Bug fixes
A print function is now available for a student to print their information into a separate txt file. 
The testing of this occurs during the start up the program, but can easily be disabled.

Some functions were bugged, and fixed. User and Course is now Comparable to make it fit to use for TreeMaps.
None of these changes should massively affect how the partially completed GUI works.